resume-nlp-service/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
spacy==3.7.4
PyMuPDF==1.23.25
python-docx==1.1.0
scikit-learn==1.4.0
pydantic==2.6.1
